Conflicts:
	data/json/mutations.json
	src/iuse.cpp

It's update time!

+++ src/trapfunc.cpp
@@ -7,7 +7,7 @@
 float pit_effectiveness(int x, int y)
 {
     int corpse_volume = 0;
-    for (int i = 0; i < g->m.i_at(x, y).size(); i++) {
+    for (size_t i = 0; i < g->m.i_at(x, y).size(); i++) {
         item &j = g->m.i_at(x, y)[i];
         if (j.type->id == "corpse") {
             corpse_volume += j.volume();
